# contracts/openapi/orqetia-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ORQETIA Client API",
    "version": "1.0.0-development",
    "description": "Canonical client/service-facing API contract. Product state: DEVELOPMENT. Backoffice/admin APIs are intentionally separate."
  },
  "servers": [
    {
      "url": "/",
      "description": "Deployment base URL"
    }
  ],
...
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/v1/sessions": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "operationId": "createSession",
        "summary": "Create an execution session.",
        "x-orqetia-required-scopes": [
          "sessions:write"
        ],
        "x-orqetia-object-ownership": "authenticated tenant/client",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created.",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/sessions/{session_id}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "operationId": "getSession",
        "summary": "Get an owned execution session.",
        "x-orqetia-required-scopes": [
          "sessions:read"
        ],
        "x-orqetia-object-ownership": "session must belong to authenticated tenant/client",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/sessions/{session_id}/tasks": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "operationId": "createTask",
        "summary": "Create asynchronous work in a session.",
        "description": "execution is optional; omission means AUTO. EXPLICIT_TARGET requires additional target authorization. The client cannot set retry/cycle/delay/timeout/quarantine/quota policy.",
        "x-orqetia-required-scopes": [
          "tasks:write"
        ],
        "x-orqetia-conditional-scopes": [
          {
            "when": "execution.mode == EXPLICIT_TARGET",
            "required": [
              "tasks:target"
            ]
          }
        ],
        "x-orqetia-object-ownership": "session must belong to authenticated tenant/client",
        "x-orqetia-content-limit-policy": "server-configured; 413 when exceeded",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Task accepted for durable asynchronous execution.",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "operationId": "getTask",
        "summary": "Get an owned task.",
        "x-orqetia-required-scopes": [
          "tasks:read"
        ],
        "x-orqetia-object-ownership": "task must belong to authenticated tenant/client",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/tasks/{task_id}/result": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "operationId": "getTaskResult",
        "summary": "Get the current/terminal task result.",
        "x-orqetia-required-scopes": [
          "tasks:read"
        ],
        "x-orqetia-object-ownership": "task must belong to authenticated tenant/client",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "operationId": "cancelTask",
        "summary": "Request idempotent task cancellation.",
        "x-orqetia-required-scopes": [
          "tasks:cancel"
        ],
        "x-orqetia-object-ownership": "task must belong to authenticated tenant/client",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "Cancellation accepted or already in equivalent cancelling/terminal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/v1/providers": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "listProviders",
        "summary": "List client-visible providers within the authenticated client's authorized envelope.",
        "x-orqetia-required-scopes": [
          "catalog:read"
        ],
        "x-orqetia-object-ownership": "result filtered to authenticated client's visible envelope",
        "responses": {
          "200": {
            "description": "Client-visible providers.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderPublicView"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or ORQETIA access credential",
        "description": "Validated according to ADR-0009/#24. Bearer may be a standards-based human/service token or an ORQETIA hashed client access credential. Required scopes are declared with x-orqetia-required-scopes."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Opaque client-generated key. Scoped by tenant/client/operation. Raw value is not a resource identity.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 200
        }
      },
...
    },
    "headers": {
      "Location": {
        "description": "Canonical resource location.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
...
      "UUID": {
        "type": "string",
        "format": "uuid"
      },
...
      "SessionCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "external_reference": {
            "type": "string",
            "maxLength": 200
          }
        }
      },
      "SessionView": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "session_id",
          "status",
          "policy_version_id",
          "created_at"
        ],
        "properties": {
          "session_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "EXPIRED",
              "CANCELLED",
              "CLOSED"
            ]
          },
          "policy_version_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
    },
...
  }
}